Added SearchableLines and SearchableLine class
Also added enums indicating the initial orientation and directions for a searchable line.
SearchableLines are to be made from the WordSearchPuzzle matrix, and are strings that span the entire row/column/diagnal axes of the word matrix. That way, the words in the WordSearchPuzzle can be easily found amongst these generated SearchableLines, with the orientation and directions dictating any reworking of the found word's indicies.